remove clauses meant for avoiding symmetries
I had code that generated clauses to avoid symmetries that would lead to
puzzle configurations that cannot be uniquely determined by any set of
hints. The generated clauses were all trivial. This could be fixed but I
doubt that this would help as there are many more source that lead to
"ambiguous" puzzle configurations.

In addition, the check for ambiguity is now done immediately after the
a puzzle configuration is generated (instead of after minimisation,
which was stupid).